+++ web-frame/src/app-shared/core/interfaces/common.interface.ts
