+++ web-frame/src/app-shared/core/interfaces/dictionary.ts
