+++ web-frame/src/app-shared/core/service/api-endpoint-builder/api-endpoint-builder.interface.ts
@@ -6,7 +6,7 @@
 
 import { InjectionToken } from '@angular/core';
 
-import { Dictionary } from '../../interface/dictionary';
+import { Dictionary } from '../../interfaces/dictionary';
 import { QueryParameterValueType, UrlParameterValueType } from '../web-request/web-request.interface';
 
 export enum ServerApi {
