+++ web-frame/src/app-shared/core/service/user/user.service.ts
@@ -8,7 +8,7 @@ import { Inject, InjectionToken } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { User } from '../../data/user.do';
-import { ICRUDService } from '../../interface/common.interface';
+import { ICRUDService } from '../../interfaces/common.interface';
 import { ServerApi } from '../api-endpoint-builder/api-endpoint-builder.interface';
 import { IWebRequestService, RequestConfiguration } from '../web-request/web-request.interface';
